android arm64-v8a arch strictness for size_t

+++ android/tangram/jni/platform_android.cpp
@@ -231,7 +231,7 @@ unsigned char* bytesFromFileSystem(const char* _path, size_t& _size) {
 
 std::string stringFromFile(const char* _path) {
 
-    unsigned int length = 0;
+    size_t length = 0;
     unsigned char* bytes = bytesFromFile(_path, length);
     std::string out(reinterpret_cast<char*>(bytes), length);
     free(bytes);
